requirements.txt
boto3 == 1.9.25
PyYAML == 3.13